requirements.txt
chromedriver-py
selenium
pytesseract 
pillow
requests
opencv-python
scipy
2captcha-python
pyyaml
openpyxl